Added a sidebar navigation for the large screen and am working on the content for the home page
