+++ lib/Pages/HomePage.dart
@@ -1,10 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:lecle_bubble_timeline/lecle_bubble_timeline.dart';
+import 'package:lecle_bubble_timeline/models/timeline_item.dart';
 import 'package:mood_swing/Pages/SignupPage.dart';
 import 'package:mood_swing/Pages/UserPage.dart';
 import '../Widgets/widgets.dart';
 import 'package:collapsible_sidebar/collapsible_sidebar.dart';
 import 'dart:math' as math show pi;
 import 'package:sidebarx/sidebarx.dart';
+
 
 import 'UserPage.dart';
 import 'LoginPage.dart';
@@ -24,165 +27,47 @@ class LargeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     double height = MediaQuery.of(context).size.height;
     double width = MediaQuery.of(context).size.width;
-    return Scaffold(
-      body: Container(
-        height: height,
-        child: SidebarLarge(),
+     return SingleChildScrollView(
+        child: Column(
+           children: [
+
+              Container(
+               width: width,
+               height: height,
+               decoration: BoxDecoration(
+               image: DecorationImage(
+                 image: AssetImage("assets/loginPageLarge.png"),
+                   fit: BoxFit.cover),
+               ),
+                child: Row(
+                  children: [
+                    Column(
+                      children: [
+                        Container(
+                          width: width,
+                          height: height,
+                          child: SidebarLarge(),
+                        ),
+                      ],
+                    ),
+                  ],
+                ),
+             ),
+
+
+
+
+
+
+
+
+           ],
+
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () {
-          ScaffoldMessenger.of(context).showSnackBar(
-            SnackBar(
-              content: Text('Yay! Button Pressed!'),
-            ),
-          );
-        },
-      ),
-    );
+     );
   }
 }
 
-///Large Sidebar
-class SidebarLarge extends StatefulWidget {
-  @override
-  _SidebarLargeState createState() => _SidebarLargeState();
-}
-
-class _SidebarLargeState extends State<SidebarLarge> {
-  late List<CollapsibleItem> _items;
-  late String _headline;
-  AssetImage _avatarImg = AssetImage('assets/music_swing_logo_small.png');
-
-  @override
-  void initState() {
-    super.initState();
-    _items = _generateItems;
-    _headline = _items.firstWhere((item) => item.isSelected).text;
-  }
-
-  List<CollapsibleItem> get _generateItems {
-    return [
-      CollapsibleItem(
-        text: 'Dashboard',
-        icon: Icons.assessment,
-        onPressed: () => setState(() => _headline = 'DashBoard'),
-        isSelected: true,
-      ),
-      CollapsibleItem(
-        text: 'Ice-Cream',
-        icon: Icons.icecream,
-        onPressed: () => setState(() => _headline = 'Errors'),
-      ),
-      CollapsibleItem(
-        text: 'Search',
-        icon: Icons.search,
-        onPressed: () => setState(() => _headline = 'Search'),
-      ),
-      CollapsibleItem(
-        text: 'Notifications',
-        icon: Icons.notifications,
-        onPressed: () => setState(() => _headline = 'Notifications'),
-      ),
-      CollapsibleItem(
-        text: 'Settings',
-        icon: Icons.settings,
-        onPressed: () => setState(() => _headline = 'Settings'),
-      ),
-      CollapsibleItem(
-        text: 'Home',
-        icon: Icons.home,
-        onPressed: () => setState(() => _headline = 'Home'),
-      ),
-      CollapsibleItem(
-        text: 'Alarm',
-        icon: Icons.access_alarm,
-        onPressed: () => setState(() => _headline = 'Alarm'),
-      ),
-      CollapsibleItem(
-        text: 'Eco',
-        icon: Icons.eco,
-        onPressed: () => setState(() => _headline = 'Eco'),
-      ),
-      CollapsibleItem(
-        text: 'Event',
-        icon: Icons.event,
-        onPressed: () => setState(() => _headline = 'Event'),
-      ),
-      CollapsibleItem(
-        text: 'Email',
-        icon: Icons.email,
-        onPressed: () => setState(() => _headline = 'Email'),
-      ),
-      CollapsibleItem(
-        text: 'Face',
-        icon: Icons.face,
-        onPressed: () => setState(() => _headline = 'Face'),
-      ),
-    ];
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var size = MediaQuery.of(context).size;
-    return SafeArea(
-      child: CollapsibleSidebar(
-        isCollapsed: MediaQuery.of(context).size.width <= 800,
-        items: _items,
-        avatarImg: _avatarImg,
-        title: 'John Smith',
-        onTitleTap: () {
-          ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(content: Text('Yay! Flutter Collapsible Sidebar!')));
-        },
-        body: _body(size, context),
-        backgroundColor: Colors.black,
-        selectedTextColor: Colors.limeAccent,
-        textStyle: TextStyle(fontSize: 15, fontStyle: FontStyle.italic),
-        titleStyle: TextStyle(
-            fontSize: 20,
-            fontStyle: FontStyle.italic,
-            fontWeight: FontWeight.bold),
-        toggleTitleStyle: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-        sidebarBoxShadow: [
-          BoxShadow(
-            color: Colors.indigo,
-            blurRadius: 20,
-            spreadRadius: 0.01,
-            offset: Offset(3, 3),
-          ),
-          BoxShadow(
-            color: Colors.green,
-            blurRadius: 50,
-            spreadRadius: 0.01,
-            offset: Offset(3, 3),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget _body(Size size, BuildContext context) {
-    return Container(
-      height: double.infinity,
-      width: double.infinity,
-      color: Colors.blueGrey[50],
-      child: Center(
-        child: Transform.rotate(
-          angle: math.pi / 2,
-          child: Transform.translate(
-            offset: Offset(-size.height * 0.3, -size.width * 0.23),
-            child: Text(
-              _headline,
-              style: Theme.of(context).textTheme.headline1,
-              overflow: TextOverflow.visible,
-              softWrap: false,
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
 
 class SmallScreen extends StatelessWidget {
   @override
@@ -190,34 +75,99 @@ class SmallScreen extends StatelessWidget {
     double height = MediaQuery.of(context).size.height;
     double width = MediaQuery.of(context).size.width;
       return Scaffold(
-        body: Stack(
-          children: <Widget>[
-            SizedBox(
-              child: Container(
-                width: width,
-                height: height,
-                decoration: BoxDecoration(
-                  image: DecorationImage(
-                      image: AssetImage("assets/ribbon_lowered.png"),
-                      fit: BoxFit.cover),
-                ),
-              ),
-            ),
-
-            SidebarSmall(),
-          ],
+        body:Container(
+        width: width,
+        height: height,
+        decoration: BoxDecoration(
+          image: DecorationImage(
+              image: AssetImage("assets/userPageLarge.png"),
+              fit: BoxFit.cover),
         ),
+          child: Stack(
+            children:
+              [
+                BubbleTimeline(
+                  bubbleSize: 100,
+                  stripColor: Colors.teal,
+                  dividerCircleColor: Colors.white,
+                  // List of Timeline Bubble Items
+                  items: [
+                    TimelineItem(
+                      title: 'Boat',
+                      subtitle: 'Travel through Oceans',
+                      icon: Icon(
+                        Icons.directions_boat,
+                        color: Colors.black,
+                      ),
+                      bubbleColor: Colors.red,
+                      description: 'Description for boat',
+                      titleStyle: TextStyle(fontSize: 25.0, fontWeight: FontWeight.w700, color: Colors.white),
+                      subtitleStyle: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w500, color: Colors.white70),
+                      descriptionStyle: TextStyle(fontSize: 12.0, color: Colors.white60),
+                    ),
+                    TimelineItem(
+                      title: 'Bike',
+                      subtitle: 'Road Trips are best',
+                      icon: Icon(
+                        Icons.directions_bike,
+                        color: Colors.black,
+                      ),
+                      bubbleColor: Colors.yellow,
+                      description: 'Description for bike',
+                      titleStyle: TextStyle(fontSize: 25.0, fontWeight: FontWeight.w700, color: Colors.white),
+                      subtitleStyle: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w500, color: Colors.white70),
+                      descriptionStyle: TextStyle(fontSize: 12.0, color: Colors.white60),
+                    ),
+                    TimelineItem(
+                      title: 'Bus',
+                      subtitle: 'I like to go with friends',
+                      icon: Icon(
+                        Icons.directions_bus,
+                        color: Colors.black,
+                      ),
+                      bubbleColor: Colors.green,
+                      description: 'Description for bus',
+                      titleStyle: TextStyle(fontSize: 25.0, fontWeight: FontWeight.w700, color: Colors.white),
+                      subtitleStyle: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w500, color: Colors.white70),
+                      descriptionStyle: TextStyle(fontSize: 12.0, color: Colors.white60),
+                    ),
 
+                    TimelineItem(
+                      title: 'Bike',
+                      subtitle: 'Road Trips are best',
+                      icon: Icon(
+                        Icons.directions_bike,
+                        color: Colors.black,
+                      ),
+                      bubbleColor: Colors.yellow,
+                      description: 'Description for bike',
+                      titleStyle: TextStyle(fontSize: 25.0, fontWeight: FontWeight.w700, color: Colors.white),
+                      subtitleStyle: TextStyle(fontSize: 16.0, fontWeight: FontWeight.w500, color: Colors.white70),
+                      descriptionStyle: TextStyle(fontSize: 12.0, color: Colors.white60),
+                    ),
+                  ],
+
+                ),
+              ],
+          ),
+
+
+
+
+      ),
     );
   }
 }
 
+
 class HomePage extends StatelessWidget {
-  static Key UserButtonKey = Key("User Button");
-  HomePage();
+ //const HomePage({super.key})
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(body: Body());
   }
 }
+
+
+
