+++ lib/Pages/SignupPage.dart
@@ -24,7 +24,7 @@ class LargeScreen extends StatelessWidget {
             height: height,
             decoration:BoxDecoration(
               image: DecorationImage(
-                  image: AssetImage("assets/LoginPageLarge.png"), fit: BoxFit.cover),
+                  image: AssetImage("assets/loginPageLarge.png"), fit: BoxFit.cover),
             ),
             child: Padding(
               ///back arrow
