+++ lib/Widgets/widgets.dart
@@ -4,3 +4,4 @@ export 'MyPalette.dart';
 export 'ResponsiveWidget.dart';
 export 'BackArrowButton.dart';
 export 'SidebarSmall.dart';
+export 'SidebarLarge.dart';
