+++ pubspec.yaml
@@ -21,6 +21,8 @@ dependencies:
   animations: ^2.0.7
   collapsible_sidebar: ^2.0.3+2
   sidebarx: ^0.14.0
+  lecle_bubble_timeline: ^0.0.1+1
+
